Fixes the old imgurpython issue (#14)
* FIX:fetchers:imgurfetcher title string encoding

The title string as obtained from the imgur library doesn't encode
characters in either ascii or utf-8, but provides a Unicode object
(which I never seem to understand). This causes a crash when trying to
log a title that contains extraneous characters (as it can be seen
with the multiplication sign under https://imgur.com/6KswIQG).

We force UTF-8 encoding on the title string to avoid this issue.

* FIX: imgurfetcher: update imgur library and logger

The old imgur library didn't work anymore, and the logger messages we
had were pretty unclear/confusing. Update the api calls and the logger
format strings to fix this.